ranking system done

+++ src/main/java/com/geomhwein/go/command/UserDetailsVO.java
@@ -19,7 +19,7 @@ public class UserDetailsVO {
 	private String userTelno; // VARCHAR(20)
 	private String userEmlAddr; // VARCHAR(100) NOT NULL
 	private Integer userAge; // INT
-	private BigDecimal userRating; // DECIMAL(5,2)
+	private Integer userRating; // DECIMAL(5,2)
 	private String gender; // ENUM('male', 'female', 'other')
 	private String address; // TEXT
 	private String profilePictureFilename; // TEXT
